Remove All-Services dashboard
The Library website dashboard is unusable because of the laggy interface
in CloudWatch. Removing this in favor of the split up dashboards. If
this gets better we can switch back.

As part of this change, added more useful descriptions of each service
to assist with understanding the impact of issues with the service. This
required adding a way to override the height of the header since these
widgets don't automatically calculate the height needed for the content.