Update to match new APIs for Library Website and more
This is mostly an update to match the new APIs, but also has some improvements
to make this code easier to maintain. Changes:
- There is no longer an Hours API, it is now more generally a Libcal Gateway
which services hours requests.
- Updated metrics and widgets to use search functions to make it easier to
maintain. As long as the lambdas and api methods follow a naming scheme
that uses the api name as a prefix, it will find the methods/lambdas
associated with the API. Any method changes or lambda changes should update
the dashboards automatically, without the need for code changes.
- Split gatekeeper slos into three: aleph, primo, and illiad gateway slos
- Dry up creation of debug widgets for an API
- An API Gateway dashboard will now include some basic latency metrics for
lambdas and methods even if there are no SLOs associated with it.
- Several cosmetic changes to debug widgets to show labels a bit better.
There was some confusion about the errors being a percentage or a count.
Making the widgets a bit larger helped with showing the labels.